% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adj_ctry}
\alias{adj_ctry}
\title{adj_ctry}
\format{
.csv
}
\source{
IIASA SSP database
}
\usage{
adj_ctry
}
\description{
File to adjust country names and homogenize differences across datasets
}
\examples{
\dontrun{
library(rhap)
rhap::adj_ctry
}
}
\keyword{datasets}
