% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adj_ctry_map}
\alias{adj_ctry_map}
\title{adj_ctry_map}
\format{
.csv
}
\source{
IIASA SSP database
}
\usage{
adj_ctry_map
}
\description{
File to adjust country names and homogenize differences between the output and the map raster
}
\examples{
\dontrun{
 library(rhap);
 rhap::adj_ctry_map
}
}
\keyword{datasets}
