% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adj_ctry_output}
\alias{adj_ctry_output}
\title{adj_ctry_output}
\format{
.csv
}
\source{
IIASA SSP database
}
\usage{
adj_ctry_output
}
\description{
File to adjust country names and homogenize differences between the output and the panel data
}
\examples{
\dontrun{
library(rhap)
rhap::adj_ctry_output
}
}
\keyword{datasets}
