% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_pollutants}
\alias{all_pollutants}
\title{all_pollutants}
\format{
.csv
}
\source{
GCAM results
}
\usage{
all_pollutants
}
\description{
List of pollutants from the residential sector
}
\examples{
\dontrun{
 library(rhap);
 rhap::all_pollutants
}
}
\keyword{datasets}
