% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hap_impacts.R
\name{calc_hap_impacts_predict}
\alias{calc_hap_impacts_predict}
\title{calc_hap_impacts_predict}
\usage{
calc_hap_impacts_predict(
  data,
  saveOutput = TRUE,
  map = FALSE,
  anim = TRUE,
  HIA_var = "deaths",
  countries = "All",
  normalized = FALSE,
  by_gr = FALSE
)
}
\arguments{
\item{data}{Dataset with ghg, gdppc and flsp estimates to compute health impacts. Structure defined
as the output of `calc_hap_impacts_preproces`}

\item{saveOutput}{Writes the emission files. By default=TRUE}

\item{map}{Produce the maps. By default = FALSE}

\item{anim}{If set to TRUE, produces multi-year animations. By default=TRUE}

\item{HIA_var}{Health metric to be predicted. c("deaths", "yll", "dalys"). By default = deaths}

\item{countries}{Target countries over which the model will be fit (defaults to \code{"All"}).
For a complete list of valid country names, check \code{unique(rhap::panel_data$country_name)}.}

\item{normalized}{Transform the output to "normalized" values. By default = FALSE}

\item{by_gr}{Estimate damages at group level. Just for illustrative purposes. By default = FALSE}
}
\value{
Health impacts attributable to HAP for all the selected years
}
\description{
Run the model to estimate the health impacts attributable to household air pollution for GCAM scenarios.
Notice that the required dataset can be obtained through GCAM or uploaded manually through the `data` parameter
}
\keyword{HAP}
\keyword{health}
\keyword{impacts,}
