% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hap_impacts.R
\name{calc_hap_impacts_preprocess}
\alias{calc_hap_impacts_preprocess}
\title{calc_hap_impacts_preprocess}
\usage{
calc_hap_impacts_preprocess(
  db_path = NULL,
  query_path = "./inst/extdata",
  db_name = NULL,
  prj_name,
  scen_name,
  queries = "queries_rhap.xml",
  final_db_year = 2100,
  HIA_var = "deaths",
  countries = "All"
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Vector names of the GCAM scenarios to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{final_db_year}{Final year in the GCAM database (this allows to process databases with user-defined "stop periods")}

\item{HIA_var}{Health metric to be predicted. c("deaths", "yll", "dalys"). By default = deaths}

\item{countries}{Target countries over which the model will be fit (defaults to \code{"All"}).
For a complete list of valid country names, check \code{unique(rhap::panel_data$country_name)}.}
}
\value{
Health impacts attributable to HAP for all the selected years
}
\description{
Prepare the data from GCAM to latter estimate the health impacts attributable to household air pollution for GCAM scenarios.
Run `calc_hap_impacts_predict` to compute the health estimates
}
\keyword{HAP}
\keyword{health}
\keyword{impacts,}
