% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hia_adder.R
\name{calc_hia_adder}
\alias{calc_hia_adder}
\title{calc_hia_adder}
\source{
IHME
}
\usage{
calc_hia_adder()
}
\value{
Adder
}
\description{
Calculate the adder for HAP impact analysis

The adder is calculated as the difference between the observed and estimated values in final data year
}
\keyword{Bias}
\keyword{adder}
