% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ResidEm_grp.R
\name{calc_ResidEm_grp}
\alias{calc_ResidEm_grp}
\title{calc_ResidEm_grp}
\usage{
calc_ResidEm_grp(
  db_path = NULL,
  query_path = "./inst/extdata",
  db_name = NULL,
  prj_name,
  scen_name,
  queries = "queries_rhap.xml",
  final_db_year = 2100,
  year,
  pollutant,
  region,
  saveOutput = TRUE,
  pie = TRUE
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Vector names of the GCAM scenarios to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{final_db_year}{Final year in the GCAM database (this allows to process databases with user-defined "stop periods")}

\item{year}{Select the year. 2020:2100}

\item{pollutant}{Select the pollutant. List of available gases in "rhap::all_pollutants"}

\item{region}{Select the region. It includes the GCAM 32 geopolitical regions + EU-27 + Global: https://github.com/JGCRI/gcam-doc/blob/gh-pages/overview.md}

\item{saveOutput}{Writes the emission files. By default=TRUE}

\item{pie}{Produce the maps. By default=FALSE}
}
\value{
Contribution of different groups to pollutant emissions
}
\description{
Generate pie-charts with the contribution of different within-region population groups to different pollutant emissions by scenario and year
}
\keyword{groups,}
\keyword{pollutants}
\keyword{population}
