% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcam_regions}
\alias{gcam_regions}
\title{gcam_regions}
\format{
.csv
}
\source{
GCAM results
}
\usage{
gcam_regions
}
\description{
List of GCAM-32 + EU-27 + Global regions
}
\examples{
\dontrun{
 library(rhap);
 rhap::gcam_regions
}
}
\keyword{datasets}
