% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_ctry.SSP1}
\alias{gdp_ctry.SSP1}
\title{GDP-SSP database}
\format{
.csv
}
\source{
IIASA SSP database
}
\usage{
gdp_ctry.SSP1
}
\description{
Filtered GDP data per SSP1 (SSP_database_2024.csv).
}
\examples{
\dontrun{
library(rhap)
rhap::gdp_ctry.SSP1
}
}
\keyword{datasets}
