% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_ctry.SSP5}
\alias{gdp_ctry.SSP5}
\title{GDP-SSP database}
\format{
.csv
}
\source{
IIASA SSP database
}
\usage{
gdp_ctry.SSP5
}
\description{
Filtered GDP data per SSP5 (SSP_database_2024.csv).
}
\examples{
\dontrun{
 library(rhap);
 rhap::gdp_ctry.SSP5
}
}
\keyword{datasets}
