% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_pc.SSP3}
\alias{gdp_pc.SSP3}
\title{GDP-SSP database}
\format{
.csv
}
\source{
IIASA SSP database
}
\usage{
gdp_pc.SSP3
}
\description{
Filtered er capita GDP data per SSP3 (SSP_database_2024.csv).
}
\examples{
\dontrun{
 library(rhap);
 rhap::gdp_pc.SSP3
}
}
\keyword{datasets}
