% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gdp_ctry.R
\name{get_gdp_ctry}
\alias{get_gdp_ctry}
\title{get_gdp_ctry}
\source{
IIASA SSP database
}
\usage{
get_gdp_ctry(ssp)
}
\arguments{
\item{ssp}{SSP scenario c("SSP1", "SSP2", "SSP3", "SSP4", "SSP5")}
}
\value{
Per capita GDP by country
}
\description{
Ancillary function to get country-level GDPpc from SSP database

SSP database version 3.0.1 updated in 2024
}
\keyword{GDPpc,}
\keyword{SSP}
