% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gdppc.R
\name{get_gdppc}
\alias{get_gdppc}
\title{get_gdppc}
\source{
IIASA SSP database
}
\usage{
get_gdppc(ssp)
}
\arguments{
\item{ssp}{SSP scenario c("SSP1", "SSP2", "SSP3", "SSP4", "SSP5")}
}
\value{
Per capita GDP by country
}
\description{
Ancillary function to get country-level GDPpc from SSP database

SSP database version 3.0.1 updated in 2024
}
\keyword{GDPpc,}
\keyword{SSP}
