% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hia_adder}
\alias{hia_adder}
\title{hia_adder}
\format{
.csv
}
\source{
Own calculations based on IHME data
}
\usage{
hia_adder
}
\description{
Bias adder, calculated as the difference between the observed and "predicted" values in final observed year
}
\examples{
\dontrun{
library(rhap)
rhap::hia_adder
}
}
\keyword{datasets}
