% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listYears}
\alias{listYears}
\title{listYears}
\usage{
listYears(projData, scenarios = NULL, queries = NULL, anyscen = TRUE)
}
\arguments{
\item{projData}{The data set to report on.}

\item{scenarios}{The name(s) of the scenario(s) to report on. If NULL, report on all of them.}

\item{queries}{The name(s) of the queries(s) to report on. If NULL, report on all of them.}

\item{anyscen}{If TRUE, then list queries that are in any scenario. If FALSE, list queries that are in all scenarios.}
}
\value{
list of years reported in the project/scenario/queries.
}
\description{
Return the years of the queries available for a scenario in a project data set.
This function requires the data set to have been previously loaded, so it cannot take a file name.
Source: gcamreport
}
