% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panel_data}
\alias{panel_data}
\title{Panel data}
\format{
.csv
}
\source{
Multiple sources (described above). These include IHME, CEDS, WB, Our World in Data, WHO, NSSO, CSY, RECS, EU-SILC, Odysee, CCK, CCSM,
}
\usage{
panel_data
}
\description{
Create a panel data to fit the econometric model for the analysis. The panel includes information of the different nations of the world and covers the period 1990-2019.

In terms of variables, the panel includes:

- Premature deaths, YLLs, and DALYs, attributable to indoor air pollution (IHME)

- Emissions from the residential sector of different pollutants, namely BC, OC, CH4, N2O, CO, NOx, NMVOC, NH3 and NMVOC. They are collected from the Community Emissions Database System (CEDS)

- Regionally-averaged ambient air pollution levels (World Bank, WB)

- Gross Domestic Product (GDP) in PPP (Our World in Data)

- Urban/Rural shares by region (WB)

- Share of population with access to electricity and clean cooking fuels (World Health Organization, WHO)

- Residential floorspace, from multiple sources: India NSSO, Chinese Statistical Yearbook (CSY), Residential Energy Consumption Survey (RECS), EU statistics on income and living conditions (EU-SILC), Odysee

- Temperature and precipitation time series  (Climate Change Knowledge Portal (CCKP))

- Heating and cooling degree days (Community Climate System Model (CCSM))
}
\examples{
\dontrun{
 library(rhap);
 rhap::panel_data
}
}
\keyword{datasets}
