% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panel_pollutants}
\alias{panel_pollutants}
\title{Selected pollutants}
\format{
vector
}
\source{
-
}
\usage{
panel_pollutants
}
\description{
Pollutants selected to be included in the ecomoetric regression model
}
\examples{
\dontrun{
 library(rhap);
 rhap::panel_pollutants
}
}
\keyword{datasets}
