% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_ctry.SSP3}
\alias{pop_ctry.SSP3}
\title{Pop-SSP database}
\format{
.csv
}
\source{
IIASA SSP database
}
\usage{
pop_ctry.SSP3
}
\description{
Filtered Population data per SSP3 (SSP_database_2024.csv).
}
\examples{
\dontrun{
 library(rhap);
 rhap::pop_ctry.SSP3
}
}
\keyword{datasets}
