% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.ssp.data}
\alias{raw.ssp.data}
\title{Raw SSP data}
\format{
.csv
}
\source{
IIASA SSP database
}
\usage{
raw.ssp.data
}
\description{
Country level population and GDP data per SSP (SSP_database_2024.csv).
}
\examples{
\dontrun{
 library(rhap);
 rhap::raw.ssp.data
}
}
\keyword{datasets}
