% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(
  db_path = NULL,
  query_path = "./inst/extdata",
  db_name = NULL,
  prj_name,
  scen_name,
  queries = "queries_rhap.xml",
  final_db_year = 2100,
  saveOutput = T,
  map = F,
  anim = T,
  HIA_var = "deaths",
  by_gr = F
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Vector names of the GCAM scenarios to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{final_db_year}{Final year in the GCAM database (this allows to process databases with user-defined "stop periods")}

\item{saveOutput}{Writes the emission files. By default=T}

\item{map}{Produce the maps. By default=F}

\item{anim}{If set to T, produces multi-year animations. By default=T}

\item{HIA_var}{Health metric to be predicted. c("deaths", "yll", "dalys"). By default = deaths}

\item{by_gr}{Estimate damages at group level. Just for illustrative purposes. By default = F}
}
\value{
Health impacts attributable to HAP for all the selected years
}
\description{
Run the model to estimate the health impacts attributable to household air pollution for GCAM scenarios
}
\keyword{HAP}
\keyword{health}
\keyword{impacts,}
