% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssp_gdp_adj}
\alias{ssp_gdp_adj}
\title{ssp_gdp_adj}
\format{
.csv
}
\source{
IIASA SSP database - SSPv9
}
\usage{
ssp_gdp_adj
}
\description{
File with GDP data for some countries that are not in the updated dataset, but i a previous version
}
\examples{
\dontrun{
library(rhap)
rhap::ssp_gdp_adj
}
}
\keyword{datasets}
