% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{fit_model}
\source{
Details on plm estimation can be found here: https://cran.r-project.org/web/packages/plm/plm.pdf#'
}
\usage{
fit_model(HIA_var)
}
\arguments{
\item{HIA_var}{Health metric to be predicted. c("deaths", "yll", "dalys")}
}
\value{
Regression model and coefficients for prediction
}
\description{
Fit the model using fixed effects model and plm function.

For the model dplyr::selection, testing and validation see "./rhap/inst/extdata/model_testing.R"
}
\keyword{Econometric}
\keyword{effects}
\keyword{fixed}
\keyword{model;}
